third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/scripts-sorting.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/debugger-test.js"></script>
<script src="../../../inspector/page-mock.js"></script>

<script>
async function test() {
  function createNavigatorView(constructor) {
    var navigatorView = new constructor();
    navigatorView.show(UI.inspectorView.element);
    return navigatorView;
  }

  sourcesNavigatorView = createNavigatorView(Sources.SourcesNavigatorView);
  contentScriptsNavigatorView = createNavigatorView(Sources.ContentScriptsNavigatorView);

  Bindings.NetworkProject.forTarget(TestRunner.mainTarget)._resetForTest();
  Bindings.resourceMapping._resetForTest(TestRunner.mainTarget);

  var pageMock = new SDKTestRunner.PageMock('http://example.com');
  pageMock.disableDOMCapability();
  SDKTestRunner.connectToPage('mock-target-1', pageMock);

  var uiSourceCodes = [];
  async function addUISourceCode(url, isContentScript) {
    pageMock.evalScript(url, '', isContentScript);
    var uiSourceCode = await TestRunner.waitForUISourceCode(url);
    uiSourceCodes.push(uiSourceCode);
  }

  function dumpScriptsList() {
    SourcesTestRunner.dumpNavigatorViewInAllModes(sourcesNavigatorView);
    SourcesTestRunner.dumpNavigatorViewInAllModes(contentScriptsNavigatorView);

    for (var i = 0; i < uiSourceCodes.length; ++i) {
      sourcesNavigatorView.revealUISourceCode(uiSourceCodes[i]);
      contentScriptsNavigatorView.revealUISourceCode(uiSourceCodes[i]);
    }

    SourcesTestRunner.dumpNavigatorViewInAllModes(sourcesNavigatorView);
    SourcesTestRunner.dumpNavigatorViewInAllModes(contentScriptsNavigatorView);
  }

  var scripts = [
    'block.js?block=foo', 'ga.js', 'lenta.ban?pg=4883&ifr=1', 'lenta.ban?pg=5309&ifr=1', 'top100.jcn?80674',
    '_js/production/motor.js?1308927432', 'i/xgemius.js', 'i/js/jquery-1.5.1.min.js', 'i/js/jquery.cookie.js',
    'foo/path/bar.js?file=bar/zzz.js', 'foo/path/foo.js?file=bar/aaa.js'
  ];
  for (var i = 0; i < scripts.length; ++i)
    await addUISourceCode('http://foo.com/' + scripts[i]);

  var scripts2 = ['foo/path/bar.js?file=bar/zzz.js', 'foo/path/foo.js?file=bar/aaa.js'];
  for (var i = 0; i < scripts2.length; ++i)
    await addUISourceCode('http://bar.com/' + scripts2[i]);
  await addUISourceCode('*Non*URL*path');

  var extensions = ['extension-schema://extension-name/bar.js', 'extension-schema://extension-name/folder/baz.js'];
  for (var i = 0; i < extensions.length; ++i)
    await addUISourceCode(extensions[i], true);
  await addUISourceCode('*Another*Non*URL*path', true);
  dumpScriptsList();
  TestRunner.completeTest();
}
</script>

</head>
<body onload="runTest()">
<p>
Tests scripts sorting in the scripts panel.
</p>
</body>

</html>
