third_party/blink/web_tests/platform/generic/http/tests/devtools/sources/debugger-ui/scripts-sorting-expected.txt

Tests scripts sorting in the scripts panel.

Sources:
-------- Setting mode: [frame]
*Non*URL*path
bar.js?file=bar/zzz.js
bar.js?file=bar/zzz.js
block.js?block=foo
foo.js?file=bar/aaa.js
foo.js?file=bar/aaa.js
ga.js
jquery-1.5.1.min.js
jquery.cookie.js
lenta.ban?pg=4883&ifr=1
lenta.ban?pg=5309&ifr=1
motor.js?1308927432
top100.jcn?80674
xgemius.js
Sources:
-------- Setting mode: [frame/domain]
(no domain)
  *Non*URL*path
bar.com
  bar.js?file=bar/zzz.js
  foo.js?file=bar/aaa.js
foo.com
  bar.js?file=bar/zzz.js
  block.js?block=foo
  foo.js?file=bar/aaa.js
  ga.js
  jquery-1.5.1.min.js
  jquery.cookie.js
  lenta.ban?pg=4883&ifr=1
  lenta.ban?pg=5309&ifr=1
  motor.js?1308927432
  top100.jcn?80674
  xgemius.js
Sources:
-------- Setting mode: [frame/domain/folder]
(no domain)
  *Non*URL*path
bar.com
  foo/path
    bar.js?file=bar/zzz.js
    foo.js?file=bar/aaa.js
foo.com
  _js/production
    motor.js?1308927432
  foo/path
    bar.js?file=bar/zzz.js
    foo.js?file=bar/aaa.js
  i
    js
      jquery-1.5.1.min.js
      jquery.cookie.js
    xgemius.js
  block.js?block=foo
  ga.js
  lenta.ban?pg=4883&ifr=1
  lenta.ban?pg=5309&ifr=1
  top100.jcn?80674
Sources:
-------- Setting mode: [domain]
(no domain)
  *Non*URL*path
bar.com
  bar.js?file=bar/zzz.js
  foo.js?file=bar/aaa.js
foo.com
  bar.js?file=bar/zzz.js
  block.js?block=foo
  foo.js?file=bar/aaa.js
  ga.js
  jquery-1.5.1.min.js
  jquery.cookie.js
  lenta.ban?pg=4883&ifr=1
  lenta.ban?pg=5309&ifr=1
  motor.js?1308927432
  top100.jcn?80674
  xgemius.js
Sources:
-------- Setting mode: [domain/folder]
(no domain)
  *Non*URL*path
bar.com
  foo/path
    bar.js?file=bar/zzz.js
    foo.js?file=bar/aaa.js
foo.com
  _js/production
    motor.js?1308927432
  foo/path
    bar.js?file=bar/zzz.js
    foo.js?file=bar/aaa.js
  i
    js
      jquery-1.5.1.min.js
      jquery.cookie.js
    xgemius.js
  block.js?block=foo
  ga.js
  lenta.ban?pg=4883&ifr=1
  lenta.ban?pg=5309&ifr=1
  top100.jcn?80674
Content Scripts:
-------- Setting mode: [frame]
*Another*Non*URL*path
bar.js
baz.js
Content Scripts:
-------- Setting mode: [frame/domain]
(no domain)
  *Another*Non*URL*path
extension-name
  bar.js
  baz.js
Content Scripts:
-------- Setting mode: [frame/domain/folder]
(no domain)
  *Another*Non*URL*path
extension-name
  folder
    baz.js
  bar.js
Content Scripts:
-------- Setting mode: [domain]
(no domain)
  *Another*Non*URL*path
extension-name
  bar.js
  baz.js
Content Scripts:
-------- Setting mode: [domain/folder]
(no domain)
  *Another*Non*URL*path
extension-name
  folder
    baz.js
  bar.js
Sources:
-------- Setting mode: [frame]
*Non*URL*path
bar.js?file=bar/zzz.js
bar.js?file=bar/zzz.js
block.js?block=foo
foo.js?file=bar/aaa.js
foo.js?file=bar/aaa.js
ga.js
jquery-1.5.1.min.js
jquery.cookie.js
lenta.ban?pg=4883&ifr=1
lenta.ban?pg=5309&ifr=1
motor.js?1308927432
top100.jcn?80674
xgemius.js
Sources:
-------- Setting mode: [frame/domain]
(no domain)
  *Non*URL*path
bar.com
  bar.js?file=bar/zzz.js
  foo.js?file=bar/aaa.js
foo.com
  bar.js?file=bar/zzz.js
  block.js?block=foo
  foo.js?file=bar/aaa.js
  ga.js
  jquery-1.5.1.min.js
  jquery.cookie.js
  lenta.ban?pg=4883&ifr=1
  lenta.ban?pg=5309&ifr=1
  motor.js?1308927432
  top100.jcn?80674
  xgemius.js
Sources:
-------- Setting mode: [frame/domain/folder]
(no domain)
  *Non*URL*path
bar.com
  foo/path
    bar.js?file=bar/zzz.js
    foo.js?file=bar/aaa.js
foo.com
  _js/production
    motor.js?1308927432
  foo/path
    bar.js?file=bar/zzz.js
    foo.js?file=bar/aaa.js
  i
    js
      jquery-1.5.1.min.js
      jquery.cookie.js
    xgemius.js
  block.js?block=foo
  ga.js
  lenta.ban?pg=4883&ifr=1
  lenta.ban?pg=5309&ifr=1
  top100.jcn?80674
Sources:
-------- Setting mode: [domain]
(no domain)
  *Non*URL*path
bar.com
  bar.js?file=bar/zzz.js
  foo.js?file=bar/aaa.js
foo.com
  bar.js?file=bar/zzz.js
  block.js?block=foo
  foo.js?file=bar/aaa.js
  ga.js
  jquery-1.5.1.min.js
  jquery.cookie.js
  lenta.ban?pg=4883&ifr=1
  lenta.ban?pg=5309&ifr=1
  motor.js?1308927432
  top100.jcn?80674
  xgemius.js
Sources:
-------- Setting mode: [domain/folder]
(no domain)
  *Non*URL*path
bar.com
  foo/path
    bar.js?file=bar/zzz.js
    foo.js?file=bar/aaa.js
foo.com
  _js/production
    motor.js?1308927432
  foo/path
    bar.js?file=bar/zzz.js
    foo.js?file=bar/aaa.js
  i
    js
      jquery-1.5.1.min.js
      jquery.cookie.js
    xgemius.js
  block.js?block=foo
  ga.js
  lenta.ban?pg=4883&ifr=1
  lenta.ban?pg=5309&ifr=1
  top100.jcn?80674
Content Scripts:
-------- Setting mode: [frame]
*Another*Non*URL*path
bar.js
baz.js
Content Scripts:
-------- Setting mode: [frame/domain]
(no domain)
  *Another*Non*URL*path
extension-name
  bar.js
  baz.js
Content Scripts:
-------- Setting mode: [frame/domain/folder]
(no domain)
  *Another*Non*URL*path
extension-name
  folder
    baz.js
  bar.js
Content Scripts:
-------- Setting mode: [domain]
(no domain)
  *Another*Non*URL*path
extension-name
  bar.js
  baz.js
Content Scripts:
-------- Setting mode: [domain/folder]
(no domain)
  *Another*Non*URL*path
extension-name
  folder
    baz.js
  bar.js
